src/test/java/com/jack2ee/webservice/domain/PostsServiceTest.java
package com.jack2ee.webservice.domain;

import com.jack2ee.webservice.domain.posts.Posts;
import com.jack2ee.webservice.domain.posts.PostsRepository;
import com.jack2ee.webservice.dto.PostsSaveRequestDto;
import com.jack2ee.webservice.service.PostsService;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest
public class PostsServiceTest {

    @Autowired
    private PostsService postsService;

    @Autowired
    private PostsRepository postsRepository;

    @AfterEach
    public void cleanup() {
        postsRepository.deleteAll();
    }

    @Test
    public void Dto데이터가_posts테이블에_저장된다() {
        // given
        PostsSaveRequestDto dto = PostsSaveRequestDto.builder()
            .author("jack2ee.dev@gmail.com")
            .content("테스트")
            .title("테스트 타이틀")
            .build();

        // when
        postsService.save(dto);

        // then
        Posts posts = postsRepository.findAll().get(0);
        assertThat(posts.getAuthor()).isEqualTo(dto.getAuthor());
        assertThat(posts.getContent()).isEqualTo(dto.getContent());
        assertThat(posts.getTitle()).isEqualTo(dto.getTitle());
    }
}
